+++ views/images/uploads.hbs
@@ -10,7 +10,7 @@
     <p>{{message}}</p>
   {{/if}}
 
-  <form action="/upload" method="POST" enctype="multipart/form-data">
+  <form action="/images" method="POST" enctype="multipart/form-data">
     <input type="file" name="image" accept="image/*" required>
     <button type="submit">Upload</button>
   </form>
